+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `get_error_counts` shows numbers of errors found in the benchmark by categories.
     - `min_counts` can return counts of both unique and non-unique min-hits with `unique_only="both"`.
     - `cross_compare` can now work with results of `compute_best`.
+    - Non-interactive scatter plots rendered by [seaborn](https://seaborn.pydata.org) are
+       created using `ResAnalyzer.seaborn_scatter_plot(tool1, tool2)`.
  * Updating files with results (all can be supplied `tool_set`)
     - `remove_tools(resfile, tools, output)` removes lines for given `tools` from `resfile`
         and stores the reduced file in `output`.